school/serializers.py
from rest_framework import serializers
from .models import Book, BookPage


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = '__all__'

class BookPageSerializer(serializers.ModelSerializer):
    book = BookSerializer(read_only=True)

    class Meta:
        model = BookPage
        fields = '__all__'