requirements.txt
# --- Django Core ---
Django>=5.0,<6.0

# --- Database (PostgreSQL) ---
psycopg2-binary>=2.9

# --- REST Framework ---
djangorestframework>=3.15.0
django-filter>=24.2

# --- JWT Authentication ---
djangorestframework-simplejwt>=5.3.1

# --- Environment & Config ---
python-dotenv>=1.0.1
django-environ>=0.11.2

# --- Static files (production ready) ---
whitenoise>=6.6.0

# --- CORS (API cross-domain) ---
django-cors-headers>=4.3.1

# --- Image / File handling ---
Pillow>=10.2.0

# --- HTTP client (useful for external APIs) ---
requests>=2.32.0

# --- Server / WSGI ---
gunicorn>=21.2.0

# --- Developer Tools (optional but recommended) ---
django-extensions>=3.2.3
pytest>=8.0.0
pytest-django>=4.8.0
black>=24.3.0
isort>=5.13.0
